Moved informers out of util package, added reference from official go/java clients
