+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/EventType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.client.utils.informers;
+package io.fabric8.kubernetes.client.informers;
 
 public enum EventType {
   ADDITION,
