+++ openshift-client/src/main/java/io/fabric8/openshift/client/DefaultOpenShiftClient.java
@@ -43,7 +43,7 @@ import io.fabric8.kubernetes.client.dsl.internal.*;
 import io.fabric8.kubernetes.client.utils.BackwardsCompatibilityInterceptor;
 import io.fabric8.kubernetes.client.utils.ImpersonatorInterceptor;
 import io.fabric8.kubernetes.client.utils.Serialization;
-import io.fabric8.kubernetes.client.utils.informers.SharedInformerFactory;
+import io.fabric8.kubernetes.client.informers.SharedInformerFactory;
 import io.fabric8.openshift.api.model.*;
 import io.fabric8.openshift.api.model.DoneableBuild;
 import io.fabric8.openshift.api.model.DoneableBuildConfig;
@@ -72,9 +72,12 @@ import okhttp3.OkHttpClient;
 import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.*;
 import java.util.Objects;
-import java.util.concurrent.Executors;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ForkJoinPool;
 
 public class DefaultOpenShiftClient extends BaseClient implements NamespacedOpenShiftClient {
 
@@ -458,7 +461,7 @@ public class DefaultOpenShiftClient extends BaseClient implements NamespacedOpen
   public SchedulingAPIGroupDSL scheduling() { return adapt(SchedulingAPIGroupClient.class); }
 
   @Override
-  public SharedInformerFactory informers() { return new SharedInformerFactory(Executors.newCachedThreadPool(), httpClient, getConfiguration()); }
+  public SharedInformerFactory informers() { return new SharedInformerFactory(ForkJoinPool.commonPool(), httpClient, getConfiguration()); }
 
   @Override
   public FunctionCallable<NamespacedOpenShiftClient> withRequestConfig(RequestConfig requestConfig) {
